+++ index.html
@@ -16,13 +16,13 @@
                 </div>
                 <div class="nav">
                     <ul>
-                        <li id="linkList" class="links">List |</li>
-                        <li id="linkAdd" class="links">Add new |</li>
+                        <li id="linkList" class="links">List&nbsp;&nbsp;&nbsp;|&nbsp;&nbsp;&nbsp;</li>
+                        <li id="linkAdd" class="links">Add new&nbsp;&nbsp;&nbsp;|&nbsp;&nbsp;&nbsp;</li>
                         <li id="linkCont" class="links">Contact</li>
                     </ul>
                 </div>  
             </div>  
-            <p class="date"></p>   
+            <div class="date"></div>   
         </header>
         <form>
             <h2>Add a new book</h2>
@@ -46,13 +46,13 @@
                 </div>
                 <div class="nav">
                     <ul>
-                        <li id="linkList" class="links">List |</li>
-                        <li id="linkAdd" class="links">Add new |</li>
+                        <li id="linkList" class="links">List&nbsp;&nbsp;&nbsp;|&nbsp;&nbsp;&nbsp;</li>
+                        <li id="linkAdd" class="links">Add new&nbsp;&nbsp;&nbsp;|&nbsp;&nbsp;&nbsp;</li>
                         <li id="linkCont" class="links">Contact</li>
                     </ul>
                 </div>  
             </div>  
-            <p class="date"></p>   
+            <div class="date"></div>   
         </header>
         <div class="books-list">
             <h2>All awesome books</h2>
@@ -72,13 +72,13 @@
                 </div>
                 <div class="nav">
                     <ul>
-                        <li id="linkList" class="links">List |</li>
-                        <li id="linkAdd" class="links">Add new |</li>
+                        <li id="linkList" class="links">List&nbsp;&nbsp;&nbsp;|&nbsp;&nbsp;&nbsp;</li>
+                        <li id="linkAdd" class="links">Add new&nbsp;&nbsp;&nbsp;|&nbsp;&nbsp;&nbsp;</li>
                         <li id="linkCont" class="links">Contact</li>
                     </ul>
                 </div>  
             </div>  
-            <p class="date"></p>   
+            <div class="date"></div>   
         </header>
         <main class="cont">
             <h2>Contact information</h2>
