--- modules/getDate.js
@@ -1,6 +1,0 @@
-const getDate = () => {
-const date = document.querySelector('.date');
-date.innerHTML = DateTime.local();
-}
-
-export default getDate;
